+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Breaking Changes
 - The ``skip_dependencies`` argument of ``build.__main__.build_package`` has been renamed to ``skip_dependency_check``
 - ``build.ConfigSettings`` has been renamed to ``build.ConfigSettingsType``
 - ``build.ProjectBuilder.build_dependencies`` to ``build.ProjectBuilder.build_system_requires``
+- ``build.ProjectBuilder.get_dependencies`` to ``build.ProjectBuilder.get_requires_for_build``